+++ stories/table.story.js
@@ -6,8 +6,8 @@ import { sorts, parseLotsToDataTable } from "../__MOCKS__/Table/helpers";
 
 const loader = {
   text: "carregando a tabelinha",
-  isEmpty: false,
-}
+  isEmpty: false
+};
 
 storiesOf("Table")
   .add("Normal", () => (
@@ -17,8 +17,21 @@ storiesOf("Table")
     />
   ))
   .add("Empty table", () => <Table translateEnum={HEADERS_TABLE} />)
-  .add("with sort", () => <Table translateEnum={HEADERS_TABLE}
-    data={parseLotsToDataTable(tableMock)} sorts={sorts} />)
-  .add("different style", () => <Table translateEnum={HEADERS_TABLE}
-    data={parseLotsToDataTable(tableMock)} striped={false} width="100%" />)
-  .add("With loader", () => <Table loading={loader} translateEnum={HEADERS_TABLE} />);
+  .add("with sort", () => (
+    <Table
+      translateEnum={HEADERS_TABLE}
+      data={parseLotsToDataTable(tableMock)}
+      sorts={sorts}
+    />
+  ))
+  .add("different style", () => (
+    <Table
+      translateEnum={HEADERS_TABLE}
+      data={parseLotsToDataTable(tableMock)}
+      striped={false}
+      width="100%"
+    />
+  ))
+  .add("With loader", () => (
+    <Table loading={loader} translateEnum={HEADERS_TABLE} />
+  ));